integrate pagination into playground
