+++ apps/gitness/src/pages/repo/repo-branch-list.tsx
@@ -8,7 +8,6 @@ import {
   PaginationItem,
   PaginationLink,
   PaginationNext,
-  PaginationEllipsis,
   PaginationPrevious,
   SearchBox,
   Spacer,
@@ -25,98 +24,6 @@ const filterOptions = [{ name: 'Filter option 1' }, { name: 'Filter option 2' },
 const sortOptions = [{ name: 'Sort option 1' }, { name: 'Sort option 2' }, { name: 'Sort option 3' }]
 
 export function ReposBranchesListPage() {
-  const PaginationComponent = ({ totalPages, previousPage, nextPage, handleClick }) => {
-    const generatePaginationItems = () => {
-      const paginationItems = []
-      const siblings = 2 // Number of adjacent pages before and after the current page
-
-      // Always show first page
-      paginationItems.push(
-        <PaginationItem key={1}>
-          <PaginationLink size="sm_icon" href="#" onClick={() => handleClick(1)} isActive={currentPage === 1}>
-            1
-          </PaginationLink>
-        </PaginationItem>
-      )
-
-      if (currentPage > 2 + siblings) {
-        paginationItems.push(
-          <PaginationItem key="start-ellipsis">
-            <PaginationEllipsis /> {/* Render ellipsis without a link */}
-          </PaginationItem>
-        ) // Ellipses before the current page
-      }
-
-      // Pages around the current page
-      for (let i = Math.max(2, currentPage - siblings); i <= Math.min(totalPages - 1, currentPage + siblings); i++) {
-        paginationItems.push(
-          <PaginationItem key={i}>
-            <PaginationLink isActive={currentPage === i} size="sm_icon" href="#" onClick={() => handleClick(i)}>
-              {i}
-            </PaginationLink>
-          </PaginationItem>
-        )
-      }
-
-      // if (currentPage < totalPages - siblings - 1) {
-      //   paginationItems.push(<span key="end-ellipsis">...</span>)
-      // }
-
-      if (currentPage < totalPages - siblings - 1) {
-        paginationItems.push(
-          <PaginationItem key="end-ellipsis">
-            <PaginationEllipsis />
-          </PaginationItem>
-        )
-      }
-
-      paginationItems.push(
-        <PaginationItem key={totalPages}>
-          <PaginationLink
-            size="sm_icon"
-            href="#"
-            onClick={() => handleClick(totalPages)}
-            isActive={currentPage === totalPages}>
-            {totalPages}
-          </PaginationLink>
-        </PaginationItem>
-      )
-
-      return paginationItems
-    }
-
-    return (
-      <ListPagination.Root>
-        <Pagination>
-          <PaginationContent>
-            {/* Previous Button */}
-            <PaginationItem>
-              <PaginationPrevious
-                size="sm"
-                href="#"
-                onClick={() => currentPage > 1 && previousPage()}
-                disabled={currentPage === 1}
-              />
-            </PaginationItem>
-
-            {/* Pagination Items */}
-            {...generatePaginationItems()}
-
-            {/* Next Button */}
-            <PaginationItem>
-              <PaginationNext
-                size="sm"
-                href="#"
-                onClick={() => currentPage < totalPages && nextPage()}
-                disabled={currentPage === totalPages}
-              />
-            </PaginationItem>
-          </PaginationContent>
-        </Pagination>
-      </ListPagination.Root>
-    )
-  }
-
   // lack of data: total branches
   // hardcoded
   const totalPages = 10
@@ -211,7 +118,7 @@ export function ReposBranchesListPage() {
       <Spacer size={5} />
       {renderContent()}
       <Spacer size={8} />
-      {/* {(brancheslistData?.length ?? 0) > 0 && (
+      {(brancheslistData?.length ?? 0) > 0 && (
         <ListPagination.Root>
           <Pagination>
             <PaginationContent>
@@ -245,8 +152,7 @@ export function ReposBranchesListPage() {
             </PaginationContent>
           </Pagination>
         </ListPagination.Root>
-      )} */}
-      <PaginationComponent totalPages={7} previousPage={previousPage} nextPage={nextPage} handleClick={handleClick} />
+      )}
     </PaddingListLayout>
   )
 }
