+++ packages/playground/src/pages/commits-list-page.tsx
@@ -4,19 +4,9 @@ import { SkeletonList } from '../components/loaders/skeleton-list'
 import { NoData } from '../components/no-data'
 import PlaygroundCommitsSettings from '../settings/commits-settings'
 import { PaddingListLayout } from '../layouts/PaddingListLayout'
-import {
-  ListActions,
-  ListPagination,
-  Pagination,
-  PaginationContent,
-  PaginationEllipsis,
-  PaginationItem,
-  PaginationLink,
-  PaginationNext,
-  PaginationPrevious,
-  Spacer,
-  Text
-} from '@harnessio/canary'
+import { PaginationComponent } from '../components/pagination'
+
+import { ListActions, Spacer, Text } from '@harnessio/canary'
 import { BranchSelector } from '../components/branch-chooser'
 import { mockRepos } from '../data/mockReposData'
 import { Link } from 'react-router-dom'
@@ -87,44 +77,13 @@ export default function CommitsListPage() {
       {renderContent()}
       <Spacer size={8} />
       {loadState === 'data-loaded' && (
-        <ListPagination.Root>
-          <Pagination>
-            <PaginationContent>
-              <PaginationItem>
-                <PaginationPrevious size="sm" href="#" />
-              </PaginationItem>
-              <PaginationItem>
-                <PaginationLink isActive size="sm_icon" href="#">
-                  1
-                </PaginationLink>
-              </PaginationItem>
-              <PaginationItem>
-                <PaginationLink size="sm_icon" href="#">
-                  2
-                </PaginationLink>
-              </PaginationItem>
-
-              <PaginationItem>
-                <PaginationLink size="sm_icon" href="#">
-                  <PaginationEllipsis />
-                </PaginationLink>
-              </PaginationItem>
-              <PaginationItem>
-                <PaginationLink size="sm_icon" href="#">
-                  4
-                </PaginationLink>
-              </PaginationItem>
-              <PaginationItem>
-                <PaginationLink size="sm_icon" href="#">
-                  5
-                </PaginationLink>
-              </PaginationItem>
-              <PaginationItem>
-                <PaginationNext size="sm" href="#" />
-              </PaginationItem>
-            </PaginationContent>
-          </Pagination>
-        </ListPagination.Root>
+        <PaginationComponent
+          totalPages={10}
+          currentPage={5}
+          nextPage={() => {}}
+          previousPage={() => {}}
+          handleClick={() => {}}
+        />
       )}
       <PlaygroundCommitsSettings loadState={loadState} setLoadState={setLoadState} />
     </PaddingListLayout>
