+++ packages/playground/src/pages/repo-webhooks-page.tsx
@@ -1,23 +1,10 @@
 import React, { useState } from 'react'
-import {
-  Spacer,
-  ListActions,
-  ListPagination,
-  Button,
-  SearchBox,
-  Pagination,
-  PaginationContent,
-  PaginationItem,
-  PaginationPrevious,
-  PaginationLink,
-  PaginationEllipsis,
-  PaginationNext,
-  Text
-} from '@harnessio/canary'
+import { Spacer, ListActions, Button, SearchBox, Text } from '@harnessio/canary'
 import { PaddingListLayout } from '../layouts/PaddingListLayout'
 import { NoData } from '../components/no-data'
 import { NoSearchResults } from '../components/no-search-results'
 import { SkeletonList } from '../components/loaders/skeleton-list'
+import { PaginationComponent } from '../components/pagination'
 import { PlaygroundListSettings } from '../settings/list-settings'
 import { mockWebhooks } from '../data/mockWebhooksData'
 import { Link } from 'react-router-dom'
@@ -87,41 +74,13 @@ function RepoWebhooksListPage() {
         {renderListContent()}
         <Spacer size={8} />
         {loadState == 'data-loaded' && (
-          <ListPagination.Root>
-            <Pagination>
-              <PaginationContent>
-                <PaginationItem>
-                  <PaginationPrevious size="sm" href="#" />
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationLink isActive size="sm_icon" href="#">
-                    1
-                  </PaginationLink>
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationLink size="sm_icon" href="#">
-                    2
-                  </PaginationLink>
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationEllipsis />
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationLink size="sm_icon" href="#">
-                    4
-                  </PaginationLink>
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationLink size="sm_icon" href="#">
-                    5
-                  </PaginationLink>
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationNext size="sm" href="#" />
-                </PaginationItem>
-              </PaginationContent>
-            </Pagination>
-          </ListPagination.Root>
+          <PaginationComponent
+            totalPages={10}
+            currentPage={5}
+            nextPage={() => {}}
+            previousPage={() => {}}
+            handleClick={() => {}}
+          />
         )}
       </PaddingListLayout>
       <PlaygroundListSettings loadState={loadState} setLoadState={setLoadState} />
