+++ packages/playground/src/pages/sandbox-executions-summary-page.tsx
@@ -1,20 +1,9 @@
 import React, { useState } from 'react'
 import { Link } from 'react-router-dom'
-import {
-  Text,
-  Spacer,
-  ListActions,
-  ListPagination,
-  SearchBox,
-  Pagination,
-  PaginationContent,
-  PaginationItem,
-  PaginationPrevious,
-  PaginationLink,
-  PaginationEllipsis,
-  PaginationNext
-} from '@harnessio/canary'
+import { Text, Spacer, ListActions, SearchBox } from '@harnessio/canary'
 import { ExecutionList } from '../components/execution-list'
+import { PaginationComponent } from '../components/pagination'
+
 import { SkeletonList } from '../components/loaders/skeleton-list'
 import { NoSearchResults } from '../components/no-search-results'
 import { NoData } from '../components/no-data'
@@ -156,41 +145,13 @@ function SandboxExecutionSummaryPage() {
         {renderListContent()}
         <Spacer size={8} />
         {loadState == 'data-loaded' && (
-          <ListPagination.Root>
-            <Pagination>
-              <PaginationContent>
-                <PaginationItem>
-                  <PaginationPrevious size="sm" href="#" />
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationLink isActive size="sm_icon" href="#">
-                    1
-                  </PaginationLink>
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationLink size="sm_icon" href="#">
-                    2
-                  </PaginationLink>
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationEllipsis />
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationLink size="sm_icon" href="#">
-                    4
-                  </PaginationLink>
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationLink size="sm_icon" href="#">
-                    5
-                  </PaginationLink>
-                </PaginationItem>
-                <PaginationItem>
-                  <PaginationNext size="sm" href="#" />
-                </PaginationItem>
-              </PaginationContent>
-            </Pagination>
-          </ListPagination.Root>
+          <PaginationComponent
+            totalPages={10}
+            currentPage={5}
+            nextPage={() => {}}
+            previousPage={() => {}}
+            handleClick={() => {}}
+          />
         )}
       </SandboxLayout.Content>
       <PlaygroundListSettings loadState={loadState} setLoadState={setLoadState} />
